:recycle: refactor: 회원기능의 uri 앞에 auth/ 추가
