+++ src/main/java/com/mulmeong/utility/adapter/in/web/controller/BookmarkController.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
-@RequestMapping("/v1/members")
+@RequestMapping("auth/v1/members")
 @RestController
 public class BookmarkController {
